refactor: latitude y longitude cambiado a Decimal

+++ app/modules/user/schemas.py
@@ -2,6 +2,7 @@ from sqlmodel import SQLModel, Field
 from pydantic import field_validator
 from app.modules.user.models import Role, UserRoleLink
 from datetime import datetime
+from decimal import Decimal
 
 
 class UserCreate(SQLModel):
@@ -43,8 +44,8 @@ class AddressRead(SQLModel):
     city: str
     province: str
     zip_code: str
-    latitud: float
-    longitude: float
+    latitud: Decimal
+    longitude: Decimal
     is_main: bool
 
     created_at: datetime
@@ -59,8 +60,8 @@ class AddressUpdate(SQLModel):
     city: str | None = None
     province: str | None = None
     zip_code: str | None = None
-    latitud: float | None = None
-    longitude: float | None = None
+    latitud: Decimal | None = None
+    longitude: Decimal | None = None
     is_main: bool | None = None
 
 
@@ -71,8 +72,10 @@ class AddressCreate(SQLModel):
     city: str = Field(max_length=100)
     province: str = Field(max_length=100)
     zip_code: str = Field(max_length=10)
-    latitud: float = Field(ge=-90.0, le=90.0, description="Latitud en grados, -90 a 90")
-    longitude: float = Field(
+    latitud: Decimal = Field(
+        ge=-90.0, le=90.0, description="Latitud en grados, -90 a 90"
+    )
+    longitude: Decimal = Field(
         ge=-180.0, le=180.0, description="Longitud en grados, -180 a 180"
     )
     is_main: bool | None = Field(default=False)
